Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -7,11 +7,11 @@
 
 int main(void)
 {
-	printf("size of a char: %d byte(s)\n", (int) sizeof(char));
-	printf("size of an int: %d byte(s)\n", (int) sizeof(int));
-	printf("size of a long int: %d byte(s)\n", (int) sizeof(long int));
-	printf("size of a long long int: %d byte(s)\n", (int) sizeof(long long int));
-	printf("size of a float: %d byte(s)\n", (int) sizeof(float));
+	printf("size of a char: %ld byte(s)\n", sizeof(char));
+	printf("size of an int: %lu byte(s)\n", sizeof(int));
+	printf("size of a long int: %lu byte(s)\n", sizeof(long int));
+	printf("size of a long long int: %lu byte(s)\n", sizeof(long long int));
+	printf("size of a float: %ld byte(s)\n", sizeof(float));
 
 	return (0);
 }
